# Conflicts:
#	lib/tools/science_and_technology/apparent_temperature/windchill/widget/windchill.dart
#	lib/tools/science_and_technology/astronomy/shadow_length/widget/shadow_length.dart
#	lib/tools/science_and_technology/telegraphs/popham_telegraph/widget/popham_telegraph.dart

+++ lib/common_widgets/dropdowns/gcw_abc_dropdown.dart
@@ -25,7 +25,7 @@ class GCWABCDropDownState extends State<GCWABCDropDown> {
       value: _currentValue ?? widget.value ?? 1,
       onChanged: (newValue) {
         setState(() {
-          _currentValue = newValue is int ? newValue as int : _currentValue;
+          _currentValue = newValue is int ? newValue : _currentValue;
           widget.onChanged(_currentValue!);
         });
       },
